Enabled gzip message compression for messages sent via the task graph.
By default this is disabled until some large scale testing and parameter studies can be performed.

To enable add this to your input file:

<Scheduler>
  <compressionLevel>5</compressionLevel>
  <compressionThreshold>50</compressionThreshold>
</Scheduler>


compressionLevel specifies the gzip compression level which ranges from 0 (no compression) to 9 (maximum compression)
compressionThreshold specifies the threshold for message size in bytes in which message compression will be attempted.  That is any messages that are of size <=compressionThreshold will not be compressed.