Fix another particle sendset bug.
The root of these problems lie in the fact that dependencies are scheduled per variable and sendsets are created once for all variables.  If you have tasks that share some variables but not all dependencies can combine in diffent ways.  This causes sendsets to be deleted even though they are still needed by other dependencies.

The current fix for this is to ensure that dependencies combine consistantly.  The current logic for combining dependences is the following:

First if a new dependency is completly included in another dependency then combine with the first one of those you find.  If no dependencies completly include the new dependency then combine it with the first dependency that can be extended to include this dependency without increasing the amount of communication. If this cannot be done then create a new dependency.

This seems to fix the sendset!=old_dw assert (for now).