When there are overlapping patches (possible with extracells) compare uda gives priority to the patch that has that cell/node in the interior.  This patch updates the tiebreaker to give priority to the patch that has the cell/node index within extra cells instead of extra nodes.
With extra nodes there is a corner case where two patches own the same node.  In this case both nodes should have the same value.  However, compare uda uses the same patch map for all variables (based on nodes).  This causes problems when looking at a cell which is on this corner case.  The current code treated the cell as a node and believed they would have the same value.  This is incorrect because with cells one patch owns the cell and the other doesn't and thus only the patch that owns the cell is gaurenteed to have the correct value.