RMCRTRadiationModel is now calling Paula's stand-alone solver.
Currently there is no passing of variables from the DW to the solver.  This is to come shortly. 

Also, I wasn't quite sure where exactly to put this task.  We might want to change it around in the future 
if we want a stand-alone solution only.  Otherwise, it may make most sense to keep it after (or even in) 
the property calculation.