M    StandAlone/sus.cc
Don't show a stack trace if a ProblemSetupException is caught.  (In
these cases, the stack trace does not provide any helpful
information.)

M    CCA/Components/ProblemSpecification/ProblemSpecReader.cc

When a tag is missing, there is not xml node from which to extract
a line number, so don't try to do it.