Combine solveEquationsMotion and integrateAcceleration into one task, computeAndIntegrateAcceleration.  This should provide a modest performance improvement, and an overall reduction in lines of code.
Also, stripped out the damping rate calculation code.  This code was fairly
indefensible and pretty much never used.  The d_dampingCoefficient flag is
still in place, and it can now used in computing the acceleration.  This is
similar to code that has been in place, but commented out, for a long time.
The damping force is now -d_dampingCoefficient*velocity.

Added a cut off mass for computing the acceleration.  If the nodal masses are
below this cut off value, then that node's acceleration will be zero.  This is
definitely a hack, but it is easy to defend based on the numerics of MPM.

These changes didn't affect the results for the MPM, MPMArches or MPMICE
regression tests.