Fixes to the code that broke when I fixed the checkForInputError() code.
M    Grid/BoundaryConditions/BoundCondFactory.cc

Clean up the Boundary Condition input code.  Not sure if we really
should use the try/catch mechanism to determine if the BC is a double
or a vector... but perhaps there isn't any other choice.  However, I
think the <BCType> tag should be fixed... perhaps add an attribute
that specifies what type it really is instead of relying on the
string in the <value> tag...?

M    ProblemSpec/ProblemSpec.h

Update the documentation on get() to be more accurate.

M    ProblemSpec/ProblemSpec.cc

* Fix indentation of #includes.
* Remove old SGI pragmas.
* use namespace std
* Fix checkForInputError() error strings.
* Use {} for if statement.