Removed the call to isSimilar from the tiled regridder.  Since the tiled regridder creates patches in predetermined locations the same functionality can be achieved using Grid::operator==() but with much less cost.  It may be worth parallelizing this operation directly in the tiled regridder.
The call to isSimilar() becomes expensive with large numbers of patches.  This could be parallelized to reduce the cost but it is now only used in the BNR regridder which is superseeded by the tiled regridder.

Added warning messages when using the BNRRegridder and the Hierarchical regridder.  These messages tell the user to use the tiled regridder which has much better performance characteristics.