Added a new SCI_DEBUG called WaitForDebugger.
This will cause sus to wait for a debugger if it crashes or throws an exception.  

To use add WaitForDebugger:+ to your SCI_DEBUG line.  When a crash occurs you will see a message that contains
hostname and the pid.  Then open a terminal on the host and attach gdb using `gdb --pid <pid>`.