initialize lp, vol and xp.
changed the variable that is used for the x axis in plotting.

Currently setup for 1 timestep and direct comparison of particle variables
with M. Stefen's 1D code.

The MMS problem runs and gives reasonable results for the particle displacement,
jowever, the particle velocity goes unstable.  We suspect that the stress
calculations needs to be moved in order to match the results of M. Steffen and 
Uintah code.