Added a couple of updates to make the script more compataible.
1)  Use readlink instead of which to ensure that the full path of the script is returned 
instead of the symbolic path.  The symbolic path is wrong after changing to the uda directory.

2) Don't open a new xterm, instead just launch the editor in the same window.  This removes the 
need for an xterm, though if EDITOR is set to an xterm editor like emacs an xwindow will still launch.
This behavior is now similar to "svn commit".

3) Use the test command instead of shell scripting on the remote host to check if a file exists.  This allows the script to work
even when the default shell script is something other than sh/tcsh. (Bash users can now rejoice).