Change the logic for determining the range that particles exist within.
Previously the computation was done using two different methods.  

The first method mapped
the particle position to a cell index and then compared that index to the low and high indices. 

The second method would map the low an dhigh indices to a position and then compare it to the 
particle position.  

In very rare cases these would not give the same answers due to floating point error causing 
a critical failure.

We now use exclusively the first method for this calculation.