First of a couple changes to how particle subsets are stored.
This removes the low and high index from the map key and changes the map to a multimap.  Queries now search this multimap and then linearlly search the resulting entries (should be between 1-7 entries) for a sufficient match.

This eliminates linear searches through the entire particle subset database in order to find a sufficient match.