BoundaryConditions.*: Standard Arches inlets can now be named. Inlet area is printed out with name when case starts up. Also allow one to set the second mixture fraction in the old ARCHES bc. This is needed for property evaluation and should be fixed at some point when the old property evaluator goes away.
InletStream.cc:
Initializing second mixture fraction. 

ExplicitSolver.cc
Tweaks on the scalar solver solve.  i
Because of the multiple density issue and the generally bad abstraction of how density is handled, the scalar eqn needs to use tempDensity to stand in for the previous rk density for the FE update. 

ScalarEqn.cc
Changes in which density is used.  
Turns out Arches is copying density from old_dw->new_dw at the beginning of each time step so old_dw density was never needed.  
Also need to use tempDensity for the explicit update.  
Finally, needed to change the "evalTransportEqn" to check if the equation used density guess to be able to handle density abstraction (see update to ExplicitSolver.cc). 

arches_ups.xml
Update spec to allow for naming of inlet and adding mixture_fraction_2 value or <FlowInlet>.