Initial commit for a particle specific load balancer.
This load balancer is attempting to address performance issues in the DLB associated with particles.

In particular, the DLB balances on a single cost metric which is a function of the number of particles 
and cells.  This load balancer seperates the metric and load balances in two phases.  During the first
phase any patches that have particles are assigned to processors and in the second phase all remaining patches
are assigned to processors.  During the first phase the load balancer attemptes to equalize the number of particles 
on each processor.  The second phase then load balances the remaining processors while attempting to equalize 
the number of cells on each processor, this is done while mainting the assignments from the first phase.

Theoretically this will help avoid synchronization problems associated with having varying numbers of particles
on each processor.