Allow variables to not be present at initialization for IO.
Previously, if a variable was created during timestepping and was requested to be output, an exception was triggered during IO for the initial condition.  This checkin changes the logic so that at the initial condition any variables that do not exist will be ignored.  After the first timestep, however, any non-existant variables that are requested to be output will trigger an exception.