First cut at getting coordinates into expression graphs in uintah.
The idea is that we calculate the fields once at initialization and then use them wherever needed by wrapping them in a placeholder expression.  However, we actually introduce a new expression here, Coordinate, which derives from Expr::PlaceHolder.  When a Coordinate expression is constructed (e.g. loaded onto an expression graph) then it will trigger Wasatch to create an expression that will populate the coordinate during initialization.

Note that this is still untested.  Tony, that is up to you!