User can now read in an auxilary file that contains two columns intensity, scalar value.
Usage: ../tools/pfs/rawToUniqueGrains [options]  <ups file> 
options:
-b, -binary:            binary output 
-l, -littleEndian:      input file contains little endian bytes  [default]
-B, -bigEndian:         input file contains big endian bytes
-auxScalarFile:         name of file that contains two columns of data (intensity scalar) 
       # case:  grains 
       # Number of blobs  75 
       # max diameter (cm) 0.0365792455209 
       # min diameter (cm) 0.00202786936452 
       # average diameter (cm) 0.00990896874237 
       # color    equivalent spherical diameter (cm) 
       1.0 0.0365792455209 
       2.0 0.0031813408649 
       3.0 0.0113385664577

The scalar value is appended to all points with that intensity.  The output files
now contain x, y, z, scalar.

It runs for a simple test problem but needs to be cleaned up and possibly templated
so the scalar can be a double, vector or whatever.