Finalize code for using Wasatch in Arches. The following changes have been made: 1. Add one way coupling between Arches fluid dynamics and Wasatch transport equations. 2. Add boundary conditions to wasatch-in-arches. All models used from Wasatch have their BCs set in Wasatch. 3. Add a Wasatch material to get things ready for MPMARCHES. 4. Wasatch in Arches now works with multipatch, multiprocessor with mpi (multipatch per processor also). 5. Multistage integrator for wasatch-arches using the Arches RK averaging interface. 6. Modify wasatch transport equations to automatically fill in the root IDs for the solution-graph-helper. 7. Add wasatch-arches input file test. This test is based on the Arches helium_1m.ups.
If you want to test Wasatch-in-Arches, there are two pieces of code in Wasatch that you must comment. This decision is based on our preference to use GhostCells instead of ExtraCells in Wasatch. When these lines of code are commented, wasatch will play well with extraCells. Please contact me or James Sutherland for additional information on what lines must be commented.