Variables with arbitrary naming conventions can now be mapped to specific CFD roles either from the input file:
<VarID>
  <var label="my_label" role="temperature"/>
<VarID>

or from individual models using ArchesLabel::setVarlabelToRole.
Varlabels are returned for specific roles from ArchesLabel using
ArchesLabel::getVarlabelByRole. This
will allow models that use the CFD quantities to easily identify needed VarLabels
if the model developer chooses not to make the specification in the
model and its input spec.

Note that permitted CFD roles are defined in ArchesLabel constructor and
should be replicated in the spec.