Added 'switcher' example using AMR.
There are three phases in this example.  In the first phase the grid and mpmice
variables are setup by running 

  inputs/UCF/Switcher/ex4_phases1_3.ups

After 3 timesteps an implicit mpm heat transfer calculation
begins by running the impm component with
  
  inputs/UCF/Switcher/ex4_phase2.ups.
  
The "simpleBurn" switch criteria states that if the reactant (matl 1) 
exceeds 350K then switch to the explosion phase (mpmice) or phase 3.