M    configure M    configure.ac
Updated the PETSc check... specifically, the previous code was using a
for loop that actually had empty parameters (but ran once with a blank
'' input).  Anyway, the logic and code was confusing, so hopefully it
is cleaner now.  The main reason for this update is now configure will
check if you are using PETSc v3.x and if so, if you specify
PETSC_ARCH, it will error out and tell you not to do this.
Previously, it would just fail without a 'good' reason.