Update to the new, non-PETSC based filter.
Switching compDynamicProcedure to use the new filter.  This will change
the answers as the new filter is not switching to a box filter at
corners, unlike the old filter. I *believe* this filter to be more
consistent.  Note that the default filter behavior near boundaries is to
renomalize the weights, like the PETSC filter.  However, one can turn
off this behavior in the input file using the "new" filter.