Adding the -pad option.  Previously, for a multi-level situation, cell indices that lay outside of the existing fine level were simply not populated with data.  This is a perfectly reasonable thing to do.  However, to enable easier post-processing, the -pad option fills those cells with zero values for the variable being extracted.  Default behavior is unchanged.
Note:  There is a big opportunity for optimization here.  Currently, for each
index for which data is wanted, every patch on the level is queried *even
after the correct patch is found*.  A factor of 2, on average, can be expected
by just quitting the search after the patch is found.  A bigger win would
come from trying that same patch that the last index was found in first.
Could be significant for petascale calculations with 10^5 patches.