Updates to the wall heat tranfer model framework, including putting it into production mode.
Removing the <const_wall_temperature> spec from the new DORadiation
model.  This was confusing and actually clashed with the
SolidWallTemperature spec in the <BoundaryConditions> section of the
input file. It is now on the user to specify BC's for the
SolidWallTemperature values if they want something other than the
resultant temperature from the indep. var space.

Updating some input files to be consistent with the new spec.

Local RT is reporting that methane_jet.ups will fail, but the
differences are in the noise.