Fixes for Kraken:
Don't use '-fPIC' if '--enable-static' is set to 'full'.  Note,
our 'static' build for the RT (due to the way linux installs)
requires some dynamic (system) libs, so we don't really force
a full static build (thus the reason we have the distinction between
the two 'static' build options).  However, on Cray machines, we
need to be fully static.  

Allow for a 'built-in' version of Hypre to be used.  Use:

  --enable-hypre=built-in