Add GPU kernel for DQMOM::constructLinearSystem. This can be further optimized by using partial products.
The idea here is based on the fact that the system of equations is normally constructed once per cell per time step, occuring on the order of several thousand times per time step. This type of high throughput computing is ideally parallelized on the GPU. In place of the normal cell-wise contruction, we can instead copy the necessary data to large arrays, indexed by cell number. We then construct all of the systems in a "single pass" on the GPU and push these results back to the host.

Also added MIT license and copyright header to DQMOM.h