*RT Update Needed*
Some corrections to:

1) Intrusion inlets: The wrong index pointer was being used in the
one-off direction from the wall.

2) Wall BC application to momentum: **This is a band-aid** Now applying
all wall bc application in BoundaryCondition::wallVelocityBC(), which
should cover domain walls and intrusions.  This sets the A and ConvA's
appropriately for the walls and for the normal velocity cells one-off
from the wall cell.  The implementation was simply to reuse Kumar's
mmbcvelocity fortran code for each coefficient set making this a little
cumbersome, but it does get the job done. We need a little cleanup to
now remove some of the defunct BC tasks that are being performed.

3) Removed scalar diffusion at inlets.  This is nothing but trouble for
mass balances.

Also, while I was at it, some minor cleanup of some unused (mms) functions.