This is an update to try and catch the problem when multiple (different) versions of zlib are linked into the same sus... This mostly has occurred on the Mac, though it seems like it could also happen under linux.  Usually, sus, one of the Uintah .so's (or .dylibs on a Mac), or a thirdparty lib that sus links against, somehow picks up a different libz.  Then, depending on which library, etc, calls for compression, potentially a different version of libz is used during the same execution.  This occurs (only?) when saving UDAs with gzip turned on, and unfortunatley is not detected until someone tries to read the UDA back in... at which point the program crashes.
I have added two fixes that will hopefully help solve this problem.

1) Makefile.in

The Makefile will now attempt to verify that sus and the Uintah
libs only link vs a single libz.  I have tested this on 'fake'
examples, but don't have a version of the code where the real problem
manifests, so am not 100% sure that it is correct.

2) Core/Grid/Variables/Variable.cc

The error manifests itself when the size of the compressed data is
read in and an invalid number is seen (usually something HUGE).  The
committed fix is a hack as we can't be sure exactly how big the
invalid number will be, but this should be ok.  Currently I am using
'1,000,000,000'.  I don't believe a single variable for a single patch
would ever be this big.  Note, the 'fix' only reports a potential
error when this occurs... though in theory if it is an error, then a
segfault will happen directly after the user notification.