Major changes to how we handle Uintah::PerPatch<double> in Wasatch.
These fields are now operable with all SpatialOps functionality including GPUs.  This is the first major change in what will likely be a few iterations to get this fully functional on GPUs, however.

The major areas of change relate to Reduction variables and any expressions that use the simulation time or time step value.