M    Point.h M    CompGeom.cc
Removed the Point's vector() operator that did a hard cast of the
Point into a Vector.  While the Point and Vector have exactly the same
data members and this cast has (in theory) worked forever, at least on
newer Macs, casting one to the other does not work.  It appears that this
was only used in one or two places, and they should just use the other
(constructor, etc) available functions for type safety reasons anyway.

Note, 'vector()' was also used in: src/CCA/Components/Wasatch/CoordHelper.cc
but James fixed this a commit or two ago.  This error was what clued us
into the problem.