Oops - forgot this file changes in an earlier commit:
It appears that there is an implementation of is_sorted within Uintah that can clash with the one in STL.  While I was tempted to rip this out, I instead just deactivated it on APPLE platforms to get us further on compiling debug builds of Uintah on the new Mac OS (Mavericks) with the latest XCode.

Framework developers: we should look into this to see if the separate implementation is really necessary.