M    DataArchiver/DataArchiver.cc
Commit a 'hack' to fix a bug in some file systems.  Some file systems
return successfully from 'mkdir' even though they have not yet
actually made the directory (I assume they have cached the call and
will take care of it whenever they feel like it).  This causes the
subsequent creation of the sub directory to fail, even though when you
go and look at the file system, the directory is there.  This is
because it 'eventually' (milliseconds later) is created, just not in
time for the making of the sub directory.

This hack puts the directory creation in a loop and will try 500 times
in a row to make the directory before it gives up and throws the exception.
During actual runs, the 2nd time 'mkdir' (createSubdir()) is called it 
works.  Also, as a note, on a 1000 processor run, it looks like only one
or two processors run into this problem (though it is never the same
processors).  

Currently there are debug statements left in the code that will report
how many times "mkdir" was called before it succeeded (if more than 1)...
Note, though, that in almost all cases, nothing is printed out as 
success happens with the 1st call.