This commit has added a regression test and is expected to cause failure in the nightly RT on coal_RadPropsPlanck.ups!
Fixed the interface for computation of absorption coefficients using the RadProps library.
The RadProps library can use one of two methods to compute the particle absorption coefficients (Planck or Rossland). 
The new R-Test is similar to the coal-channel case called coal_RadPropsPlanck.

This commit also removed a work-around (or hack) due to a boundary condition bug, which was fixed by Tony Saad.