This fixes the lineextract problem seen on BGQ machine vulcan@llnl. However, it is likely that this same problem will/does occur on any strange statically linked architecture.  While this isn't a complete fix, it will allow Jennifer to use the utility for now.
The problem in this case is that static variables are not always
(ever?) created (or at least their constructors are not called) upon
program startup (like they are supposed to be).  The Uintah type
system creates a list of all variable types (eg: CCVariable<double>)
that are available to the program, and these are stored in the static
map variables 'types' found in Core/Disclosure/TypeDescription.cc.
However, these types are put into the 'types' variable by the
register_type() function which is called from the TypeDescription
constructor for each type.

Later, when a (textual) type decription is read from an UDA file, that
string is looked up in the 'types' map... If the map was never populated,
then an 'unknown type' type is returned and then an error occurs because
we don't know the real type.

To solve this problem, I just manually instantiate a number of
variable types in the bgq_hack() function in lineextract.cc.

M    StandAlone/tools/extractors/lineextract.cc

- Line up variables (and their initialization) for easier viewing.
- Add white space, remove tabs and indent properly.
- Number of coding standard stylistic updates.
- Added bgq_hack() function to manually instantiate needed types.
- Added in warning message in case a needed type hasn't been hacked in yet.

M    StandAlone/tools/extractors/timeextract.cc

- White space fixes.
- This file has not been 'hacked' like lineextract, but if it is needed, the
     same hack would be easy to copy over.

M    Core/DataArchive/DataArchive.cc
M    Core/Disclosure/TypeDescription.cc

- Cosmetic updates - white space and coding standard clean ups.