M    Makefile.in M    Core/Util/sub.mk
Fix a strange linking error on Prism@CHPC (linux box) where the
tracker program would not link correctly due to missing symbols in
Core/Util (which are found in Core/Malloc) when SCI Malloc is turned
on.  While this fix is valid (Core/Util does rely on Core/Malloc), it
should not be necessary as Core/Malloc is linked in other locations.