RMCRT should be using the OLDDW values of abskg.
This was fixed by carrying forward the radiation properties from the
previous timestep rather than changing the DW in the RMCRT source term
due to a grid issue with the dataonion approach.

Also, fixed a missing call to commonProblemSetup for the radiation
properties.  This was causing properties to have bogus values on the
first timestep that could lead to non-terminating rays.

This commit may change the answers.