Delay deletion of Ray object.
There seems to be a race condition surrounding the deletion of the SimulationInterface object (UintahParallelComponent) in sus.cc. When this is done, it appears that for the RMCRT_Test destructor may evaluate d_RMCRT as non-null before it is deleted. This allows a second attempt to delete d_RMCRT, causing a double delete.

NOTE: credit goes to John Holmen for finding this one. Thanks John.