Tested on Ash, BigMac, and Aurora.
Fixes for Tony S. so that new routines can be placed in SolverInterface.cc.

Here is the fix for placing code from SolverInterface.h into SolverInterface.cc.
The two issues fixed with this commit are 1) Added SolverInterface.cc to the 
list of files to compile, and 2) added Core/Math to the list of
libs that must be linked (as it was missing), both in sub.mk

M    Ports/SolverInterface.cc
M    Ports/SolverInterface.h

- Moved some new functions from the .h file into the .cc file. 
- Align, remove tabs, clean up.

M    Ports/sub.mk

- Add SolverInterface.cc to list of files to compile.
- Align, remove extra spaces, alphabetize, make eaiser to follow, remove duplicate lib listings.
- Added Core/Math to the list of libs to link against.