The beginnings of a MultiScaleSimulationController.
This will be in lieu of hacking around the AMR infrastructure, resulting in likely negative side effects for AMRMPM, etc.

For now we'll explore the doMDOnLevel() route for MD on a separate level from MPM and the associated particle relocation issues, but hope to explore a more robust solution to this in the not-too-distant future.


Former-commit-id: 0df0a8fabf068c6b985c2f686f670ee5b227ed36