The beginnings of a MultiScaleSimulationController.
This will be in lieu of hacking around the AMR infrastructure, resulting in likely negative side effects for AMRMPM, etc.

For now we'll explore the doMDOnLevel() route for MD on a separate level from MPM and the associated particle relocation issues, but hope to explore a more robust solution to this in the not-too-distant future.


Former-commit-id: f030e9b9f3d40bb107f76d5e1d39aeab20ce5605