This commit changes the answers. The answer changes are due to changes in the dt computed in computeStableTimestep. The code now uses patch->dCell for grid spacing rather than cellInfo (which should be removed completely from Arches).
**bigger change**
The solver abstraction in Arches has been cleaned up. As a result,
all the solver information needed for the ExplicitSolver has been moved
into itself. This should contain the historic solver in Arches without
stepping on any newly abstracted solvers. This last change should not
affect the answers as it only moved problemSetups and allocation of
variables into the solver itself. Ordering was preserved.