-- Implement a levelBased doInitialTimestep for the MultiScaleSimulationController.
This should be totally transparent to any non-multiscale component on this branch, but the rabbit hole was pretty deep for this stuff!

Currently the main changes are couched in:

MultiScaleSimulationController::doLevelSetBasedInitialTimestep(...)

This function is not called even in the MultiScaleSimulationController since it's in heavy flux.   However, nobody should be expecting MultiScaleSimulationControllers to work properly at the moment; every single component ought to call a different controller and component, so development of other portions of this branch should, again, not be affected.


Former-commit-id: 5b387ec7d04c19663e96d8208df3b88977e3c6be