Redirect all MPI calls through the lightweight wrapper (header only)
This allows for error checking and easy collection of runtime stats.
All MPI funtions can be called from the wrapper by replacing
MPI_ with MPI::
  ex
  MPI::Isend(...)
  MPI::Reduce(...)

I checked this against truck, since rev 55071 broke builds on the mac
and a later rev is creating seg faults in the mpi scheduler on linux.


Former-commit-id: af2cdf7357a3fbcced8f6fdaec0fdf83bcb55f0e