modUPS.py
added modUPS2() which allows the user to modify xml tags that appear more than once in a ups file.  For example:

RMCRT_DO_perf_GPU_ups = modUPS2( the_dir, \
                               "RMCRT_DO_perf.ups", \
                               ["/Uintah_specification/Grid/Level/Box[@label=0]/resolution :[32,32,32]",
                                "/Uintah_specification/Grid/Level/Box[@label=0]/patches    :[2,2,2]",
                                "/Uintah_specification/Grid/Level/Box[@label=1]/resolution :[64,64,64]",
                                "/Uintah_specification/Grid/Level/Box[@label=1]/patches    :[4,4,4]",
                                "Uintah_specification/RMCRT/nDivQRays                      : 100"
                               ] )
                               
You could also use it to modify boundary conditons.                          

I'm not 100% happy with this solution since it depends on (xmlstarlet), which isn't in the standard linux/OSX.  It's trival to install
on the debian/ubunbu OSs.  If it's not installed it will warn the user and move on.

Examples.py:

Added RMCRT performance tests 1L and DO for the GPU.