* Use variadic templates to simplify and generalize Action and Task constructors.
* Added utility class (header-only), src/Core/Util/TupleHelpers.hpp, that provides a clean way to deal with template or function parameter packs.

To see the utility in the amount of code reduction this achieves, compare Task.h now and with the previous revision. We have one Action and Task CTOR for CPU and GPU, instead of 5 overloads (for each, CPU adn GPU) and can now take in an arbitrarty number of additional arguments.

Refactor Task class and supporting classes:
  * member naming convention - m_
  * use nullptr comparison/checks over checks for 0
  * heavy cleanup and formatting throughout