A temporary hack to get around const correctness enforced by the new MPI wrappers. This is ONLY for the old RT/buildbot machine, due to the ancient MPI installations on all the VMs, etc.
NOTE: This quick fix is temporary, and will be removed once the new RT/buildbot machine is online, which should be soon. We simply need to maintain continuous integration and testing until that time. If we need, we will update the MPI on the VMs (e.g. the new RT machine is not online soon enough).