Allow ghost cell requirements to vary across levels for tasks scheduled on the fine level, e.g. RMCRT_DO. This reduces the LoadBalancer::createNeighborhood complexity for RMCRT_DO significantly, e.g. from 45 minutes to 15 seconds with 64^3 patches on the fine level.
NOTE: This is a unique case and the solution was a trade off: a simple design for the edge-case far outweighed the complexity a generalized solution involved. This was effectively the addition of one integer in Task and SchedulerCommon and some simple logic elsewhere.

Also fixed some logic on sched_rayTrace_dataOnion to more clearly cycle through course levels for requriements.