***This will change answers***
Fixing a bug that was introduced when attempting to set tangential
velocity BCs to be consistent on the face of the computational domain.
The index was moving in the wrong direction at the boundaries which
resulted in tangential velocities being set on the interior flow cell
next to the boundary rather than in the extra cell.