Fix a nasty synchronization bug in checkpointing based on wall clock time. This, I believe, mainly only shows up with very quick time steps, but it could, conceivably, show up 'randomly' any time.  The problem was that each rank was determining "what the wall clock time is"... And each processors clock could be off, or the check could be just a micro second delayed, causing one core to not think that it is time for a checkpoint, and all others to try to checkpoint, thus causing a hang (or segfault).
It would appear that this issue has been in the code for a long time.
In almost (?) all cases where real time is used, the time is found by
rank 0 and broadcast to everyone else.  However, (at least) in this case
all ranks were calling currentSeconds().  I believe that I have found
the two locations this was happening and fixed it, and want to get this
fix committed for Sid to try and so that Ben's runs can be attempted.

However, it would behoove us to actually go through this code in detail
to make sure that we haven't missed anything and that redundant work
is not being performed.  Also, I don't think that the broadcast will
be a major problem, but I'd like to get Alan's (et al) thoughts on this.

M    CCA/Components/DataArchiver/DataArchiver.cc

- Removed tabs, Minor white space clean ups. Use > 80 columns of code where it makes sense.
    Be consistent with placement of '{' in for loops, and use of else on a new line.
- d_checkpointWalltimeInterval is an integer, though there were (are?) a
    few places it was being treated like a float.  Probably didn't make a
    difference, but we should be consistent.
- Have rank 0 determine the time and bcast it if we are doing wall clock
    check pointing.

M    CCA/Components/SimulationController/AMRSimulationController.cc

- Have rank 0 determine the time and bcast it if we are doing wall clock
    check pointing.