-- Clean up the SE Materials input directory.
For now, directories named AlternateInputFiles, AlternateMaterialFiles, and Utilities have been created to house the older versions.  These should be treated as archive directories and will be removed once the reactive code is backported and the necessary materials for reactive test cases (as well as the reactive/phase change test cases) have been re-established.