Previously we used a VarHashMap to keep track on our variables and where they were located in the DataArchive (UDA) files on disk. However, as it was a hash map, there was no prescribed order to how the variables were stored in that map.  When using PIDX, the variables must be read and written in exactly the same order.  To solve this problem, instead of hashing, I have just created a vector to hold this data.  The change is fairly straightforward, and Todd and Sid have tested this, and my own local RT over ICE works, so I am fairly certaint that this will not break the RT.
However, I have not analyzed how often this code is actually called
during a simulation and there could be some performance issues.
Though I don't think this should be the case, I am leaving this note
in the commit log in case someone sees an issue in the future
hopefully it will make it easier to track down.  The performance issue
would come from the fact that instead of pulling a variable directly
out of the hash map [O(1)], we now have to search the vector [O(n)]
for it.  If this is shown to be an issue, we will need to rethink how
to solve the two issues (savinging in a defined order - while not
having to linearly search the vector). -Dav