* Added Damodar's parallel Hypre solve prototype. * Cleaned up whitespace and formatting.
This commit allows the Hypre solve task to be executed in parallel within an MPI process. This is accomplished by splitting the PerProcessor PatchSet across the desired number of OpenMP threads and parallelizing the solve over the resulting per-thread PatchSets. For serial execution within an MPI process, the original PerProcessor Patch set is used with a single iteration of the solve loop.

Note, this functionality requires a modified version of Hypre to use multiple threads within an MPI process. Using these modifications, this functionality can be invoked by defining the MODDED_HYPRE macro and specifying the number of OpenMP threads to use with the HYPRE_THREADS environment variable.


Former-commit-id: fb5d4538f851a5bbdde03392e9072421ce2d16a9