* Added a tarball featuring Damodar's initial changes to hypre 2.11.2.
This commit adds the modified version of hypre referenced in commit r57387 and features changes to the files below:

* hypre-2.11.2/src/utilities/_hypre_utilities.h
* hypre-2.11.2/src/utilities/mpistubs.c

These changes allow hypre to use multiple threads within an MPI process. Using these modifications, this functionality can be invoked by defining the MODDED_HYPRE macro and specifying the number of OpenMP threads to use with the HYPRE_THREADS environment variable.


Former-commit-id: e35267f5d3044ef2f2c06a256a1556b38bb78abc