* Added Damodar's parallel Hypre solve prototype. * Added a tarball featuring Damodar's initial changes to hypre 2.11.2. * Cleaned up whitespace and formatting.
This commit merges code from the komp_scheduler branch with the kokkos_dev branch to allow the hypre solve task to be executed in parallel within an MPI process. This is accomplished by splitting the PerProcessor PatchSet across the desired number of OpenMP threads and parallelizing the solve over the resulting per-thread PatchSets. For serial execution within an MPI process, the original PerProcessor Patch set is used with a single iteration of the solve loop.

Note, this functionality requires the modified version of hypre included with this commit, which features changes to the files below:

* hypre-2.11.2/src/utilities/_hypre_utilities.h
* hypre-2.11.2/src/utilities/mpistubs.c

Using these modifications, this functionality can be invoked by defining the MODDED_HYPRE macro and specifying the number of OpenMP threads to use with the HYPRE_THREADS environment variable.


Former-commit-id: 2e550eef46d2e95b056ca07b946a9dec8199758d