1. Jeremy's commits for WBCHelper.cc and SurfaceVolumeFractionCalc.cc
Fixing the staggered volume fractions. This will change the number of
variables in the checkpoint which causes multiple failures despite the
answers being the same.

As part of this commit, had to remove an edge cell addition to the
Uintah boundary iterators which was causing edge iterators to be added
to non-adjoining geometry.

2. Stress tensor: add eps multiplication to gradients

3. Move BC for A matrix to initialization


Former-commit-id: 7e749aa4b13e78eacc9e72dc07145990ecff0361