Added test for dynamic radiation solving frequencies.
This commit also constrains the solve to be within the solve frequency and the timescale estimated at runtime.
This is the first test to use perpatch variables in the checkpoint, so compare_uda will need to be extended to support this data type (and possibley PIDX? yuck)
Wall model integration is forthcoming.



Former-commit-id: 5239dc5d38c7047cc054d644c5af8647bcc897ea