* Added KokkosScheduler.cc/h. * Added support for scheduler type "Kokkos".
This commit creates a new scheduler that will be designed primarily to support simultaneous use of the Kokkos::OpenMP and Kokkos::CUDA back-ends.

The scheduler created with this commit is a snapshot of UnifiedScheduler.cc/h at kokkos_dev commit r59223 and differs only in name.

This scheduler can be invoked explicitly from within an input using the below:

<Scheduler type="Kokkos">
</Scheduler>


Former-commit-id: 61f8f6ccf08c723a7110c00c680654f3de9edfaa