Fix a sign error in the intrusion inlet when the inlet is the in the negative direction.
Clean up the pruning of intrusions across patches so that the geometry
bounding box is used. Somehow this was changed to the patch box in an
old commit.

Clean up documentation in the arches_spec and require that the massflow
be specified as a positive value for all faces. The sign is now assigned
based on the face normal.


Former-commit-id: 4e70aa9eab59cd33a5028b762a250d96b5a3cb44