Fixed crash in almgren-mmsBC.ups and isotropic_kokkos_wale.ups.
Fix in revision 59935 introduced D2H copying of neighboring patches so that neighboring patches will be available when getGridVar is called by the task. But sometimes list of neighbors contains patches outside the *domain*, which are non existent. This happens due to calculating neighboring patches on GPU without considering validity of the variable on cpu. Added check to ignore such patches in D2H copy.


Former-commit-id: 8b8f17b16d19902b8a7a435464d4ba26edf66afc