Temporary fix for another race condition in UnifiedScheduler::initiateH2DCopies and UnifiedScheduler::prepareDeviceVars:
Race condition: Threads 0 and 1 run tasks A and B and need variable X
with ghost cells 0 and 1, respectively. Note the number ghost cells is
overridden in task graph only if it is greater than 0. So this special
case does exist. Now the two threads enter initiateH2DCopies at the same
time and both detect gpu status of X as not valid and not copying in. So
both the threads go into "else if (( !allocated ) || ( allocated &&
correctSize && !validOnGPU && !copyingIn )) {" and aim to gather ghost
cells on the host and then transfer everything to device. Later when
prepareDeviceVars is called, only one thread succeeds in getting
"performCopy" flag (compareandswap in gpudw). If thread 0 succeeds it
does not gather ghost cells and thread 1 does not copy anything. Later
task B produces wrong results as ghost layer contains garbage values.

Temporary fix: Changed the condition "if(performCopy)" to
"if(performCopy) || numGhostCell > 0" in
UnifiedScheduler::prepareDeviceVars. This will ALWAYS carry H2D transfer
if ghost cells are greater than 0. Thus extra overhead of redundant
copying for now, but makes heliumKS_pressureBC work. Long term fix can
be to repeat the copy only if current copy has num of ghost cells set to
0.
This should not cause a problem if the variable is being modified by
another GPU task concurrently, because ideally scheduler should execute
such tasks "exclusively". But needs to be tested. No such task in the
mix as of now.


Two more small fixes: Commented dtask->clearPreparationCollections() in
allHostVarsProcessingReady and allGPUVarsProcessingReady. Thread
sanitizer reported race conditions for this clearing. It does not affect
the logic. Added dwIndex check in allGPUVarsProcessingReady to ensure
duplicate entry in map if the dw is different. This is for the tasks
that need same variable from old and new dw.