Fixing max ghost cells across init and main task graphs:
Resizing GPU variables is no longer a problem since the fix in
UnifiedScheduler allocates variables with max ghost cell values.
The max ghost cells values for variables across tasks are computed in
SchedulerCommon::addTask, which is invoked during schedulerInit,
scheduleTimestep etc. But still the problem occurs sometimes especially
while using multiple taskgraphs. e.g., Consider two taskgraphs: init and
main. A GPU task in init task graph computes a variable on GPU, but is
not use by any other task in init. So the max ghost cells computed
during init task graph is 0 and the GPU memory is allocated.
Now a GPU task in the main task graph uses the same variable with ghost
cell 1, but the allocation was already done with ghost cells = 0 and we
get this error. Hence here we call scheduleTimeAdvance etc.to look up
future tasks and compute max ghost values across init and main task
graphs and seems to work for Arches.
 **** Few possible problems: ****
1. The fix seems to work for Arches, but still may fail while using sub
scheduler such as Poisson2.
2. The fix is not tested with components other than Arches and also for
AMR problems and might fail. As of now calling only scheduleTimeAdvance,
scheduleFinalizeTimestep and scheduleComputeStableTimeStep. We might
have to call more functions depending on the need.
3. collectGhostCells MUST be called AFTER init task graph in compiled,
but BEFORE task graph is executed. Some tasks in main task graph of
Archesare dependent on some setup done during compilation of init task
graph (WBCHelper of Arches), so scheduleTimeAdvance must be called after
scheduleInit. But also we dont want to execute main tasks during
initialization. So call collectGhostCells after init task graph is
compiled. That way, even if new tasks are added in task graph,
corresponding DetailedTasks are not created and not executed when
scheduler->execute() is called. Also it is important    to call
collectGhostCells before executing init tasks so that the scheduler gets
the up to date information about ghost cells. If the logic causes
problem in future, we might have to explicitly delete extra tasks added.

Added function AMRSimulationController::collectGhostCells  and calling
it in doInitialTimestep.